requirement.txt
autopep8==1.6.0
beautifulsoup4==4.10.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.6.0
idna==3.3
jieba3k==0.35.1
joblib==1.1.0
lxml==4.8.0
mypy-extensions==0.4.3
newspaper3k==0.2.8
nltk==3.7
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.1
pycodestyle==2.8.0
python-dateutil==2.8.2
python-docx==0.8.11
PyYAML==6.0
regex==2022.3.2
requests==2.27.1
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
tinysegmenter==0.3
tldextract==3.2.0
toml==0.10.2
tomli==2.0.1
tqdm==4.63.0
typing-extensions==4.1.1
urllib3==1.26.8
